requirements.txt
langchain
langchain-openai
markdownify
python-dotenv
beautifulsoup4
lxml
requests